Data Project [Part 1 of 3]
Web Scraping with REQUESTS library;
Analysis with BeautifulSoup and Pandas;
Use of SELENIUM for advanced Web Scraping, initializing the Web Driver, executing the Java Script and getting the rendered HTML;
Combining all data into a single Pandas data frame and writing it to CSV for players, teams and mvp's;
All HTML files archived and saved (in case we want to go back and process them in a different way)
Next steps: Prepare the data for Machine Learning. So getting the predictors right and making sure you clean the data properly.